backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
fairlearn==0.10.0
scikit-learn==1.3.2
google-cloud-storage==2.14.0
google-cloud-firestore==2.13.0
google-cloud-aiplatform==1.41.0
python-dotenv==1.0.0
pydantic==2.5.0
reportlab==4.0.8
pillow==10.1.0
openpyxl==3.1.2
requests==2.31.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0